+++ CISProject/src/businessLogicLayer/ExceptionHandler.java
@@ -4,12 +4,32 @@ import businessLogicLayer.Account;
 
 public class ExceptionHandler extends Exception{
 	
+	
 	public ExceptionHandler(String message) {
 		super(message);
-
 	}
-}
 	
+	//this method is called when a new user tries to register an account but before it is stored in the database
+	//the purpose is to check the format of their input (i.e.) email address, SSN, etc...
+	public void checkRegistration(Account account) {
+		
+		try {
+		
+		String ssnString = "" + account.getSsn();		
+		
+		int lengthOfSsn = String.valueOf(ssnString).length();
+		
+		if(lengthOfSsn != 9) {
+			throw new SSNException("Something wrong with the social security number");
+		}
+		}catch(SSNException ex) {
+		ex.toString();
+		}
+		
+		
+	}
+
+		}
 
 	class LoginException extends ExceptionHandler{
 		
@@ -44,35 +64,13 @@ public class ExceptionHandler extends Exception{
 		}
 		
 	}
-	class SSNException extends Exception{
+	class SSNException extends ExceptionHandler{
 		
 		private static final long serialVersionUID = 1L;
 
 		public SSNException(String message) {
 			
 			super(message);
-		}
-		
-		
-		//this method is called when a new user tries to register an account but before it is stored in the database
-		//the purpose is to check the format of their input (i.e.) email address, SSN, etc...
-		public static void checkRegistration(Account account) throws SSNException {
-			
-			
-			try {
-			
-			String ssnString = "" + account.getSsn();		
-			
-			int lengthOfSsn = String.valueOf(ssnString).length();
-			
-			if(lengthOfSsn != 9) {
-				throw new SSNException("Social Security No. must be 9 numbers");
-			}
-			}catch(SSNException e) {
-				e.getMessage();
-				
-		}
-		
 		}
 	}
 	
@@ -94,7 +92,6 @@ public class ExceptionHandler extends Exception{
 		
 			super(message);
 		}
-		
 	}
 	 class UserNameException extends ExceptionHandler{
 		
@@ -141,7 +138,3 @@ public class ExceptionHandler extends Exception{
 		}
 		
 	}
-	
-
-		
-
