+++ CISProject/src/businessLogicLayer/Flight.java
@@ -2,7 +2,7 @@ package businessLogicLayer;
 
 import java.util.ArrayList;
 import businessLogicLayer.Account;
-import java.util.Random;
+
 
 public class Flight implements Comparable<Flight> {
 	
@@ -18,11 +18,11 @@ public class Flight implements Comparable<Flight> {
 	public Flight() {
 
 	}
-		//Creates a flight and sets a random flight number
-	public Flight(String departureCity, String destinationCity, String departTime, String arriveTime, 
-			String flightDate,int numberOfSeats) {
+
+	public Flight(String departureCity, String destinationCity, String departTime, String arriveTime, String flightDate,
+			int flightNumber,int numberOfSeats) {
 		
-		setFlightNumber();
+		this.flightNumber = flightNumber;
 		this.departureCity = departureCity;
 		this.destinationCity = destinationCity;
 		this.departTime = departTime;
@@ -30,11 +30,6 @@ public class Flight implements Comparable<Flight> {
 		this.flightDate = flightDate;
 		this.numberOfSeats = numberOfSeats;
 
-	}
-	//called from GUI to create flight object and insert to database
-	public static void generateFlight() {
-		
-		
 	}
 	
 	public String getPassengers(ArrayList<Account> passengers) {
@@ -51,14 +46,6 @@ public class Flight implements Comparable<Flight> {
 		}
 			
 		return toPrint;
-	}
-	public void setFlightNumber() {
-		
-		Random rand = new Random();
-		
-		int ID = rand.nextInt(9999);
-
-		this.flightNumber = ID;
 	}
 	
 	public void bookFlight(Account account) {
