+++ CISProject/src/graphicUserInterface/NightSky.png
